fix(lint): split mode-switch constants out of component modules
