+++ client/dashboard/src/components/mode-switch-stage.tsx
@@ -1,118 +1,27 @@
-// oxlint-disable react/only-export-components -- provider + surface + hook ship together
 import {
   BRAND_MESH_SURFACE_CLASS,
   BrandMeshLayers,
 } from "@/components/brand-mesh";
 import { Icon } from "@/components/ui/Icon";
-import type { IconName } from "@/components/ui/Icon/names";
 import { cn } from "@/lib/utils";
-import {
-  createContext,
-  useCallback,
-  useContext,
-  useEffect,
-  useRef,
-  useState,
-} from "react";
 import { HeadlessContent } from "@/pages/org/HeadlessContent";
+import { useCallback, useEffect, useRef, useState } from "react";
 import { useNavigate } from "react-router";
-
-export type Mode = "canvas" | "headless";
-
-// Tab order in the switcher, which is also left-to-right slot order in the
-// grid the panes shrink into.
-export const MODES: Array<{ mode: Mode; label: string; icon: IconName }> = [
-  { mode: "canvas", label: "Dashboard", icon: "square-mouse-pointer" },
-  { mode: "headless", label: "Headless", icon: "terminal" },
-];
-
-const slotOf = (mode: Mode) => MODES.findIndex((entry) => entry.mode === mode);
-
-// Chrome-for-iOS tab switching, in three beats: the live pane shrinks into its
-// card slot, both cards sit side by side for a moment, then the chosen card
-// zooms back up to fill the pane.
-const SHRINK_MS = 520;
-const HOLD_MS = 380;
-const ZOOM_MS = 560;
-const EASE_OUT = "cubic-bezier(0.32, 0.72, 0, 1)";
-
-// Height of the mode strip above the panes — the grid is laid out underneath
-// it. Kept in sync with MODE_SWITCHER_HEIGHT in mode-switcher.tsx.
-const STRIP_HEIGHT_PX = 48;
-const CARD_GAP_PX = 24;
-const GRID_MAX_WIDTH_PX = 1200;
-const GRID_INSET_PX = 96;
-
-type CardGeometry = {
-  transform: string;
-  left: number;
-  top: number;
-  width: number;
-  height: number;
-};
-
-type Grid = {
-  cards: [CardGeometry, CardGeometry];
-  paneWidth: number;
-  paneHeight: number;
-  scale: number;
-};
-
-/**
- * Where each mode's card sits, and the transform that maps a full-size pane
- * onto it. The pane spans the viewport below the strip, so its rect is derived
- * from the window rather than measured — the panes are mid-animation when this
- * is read, and a measured rect would already include the transform.
- */
-function computeGrid(): Grid {
-  const paneWidth = window.innerWidth;
-  const paneHeight = window.innerHeight - STRIP_HEIGHT_PX;
-  const available = Math.min(paneWidth - GRID_INSET_PX, GRID_MAX_WIDTH_PX);
-  const cardWidth = (available - CARD_GAP_PX) / 2;
-  const scale = cardWidth / paneWidth;
-  const cardHeight = paneHeight * scale;
-  const originLeft = (paneWidth - (cardWidth * 2 + CARD_GAP_PX)) / 2;
-  const top = STRIP_HEIGHT_PX + (paneHeight - cardHeight) / 2;
-
-  const cards = [0, 1].map((index) => {
-    const left = originLeft + index * (cardWidth + CARD_GAP_PX);
-    return {
-      left,
-      top,
-      width: cardWidth,
-      height: cardHeight,
-      // transform-origin is the pane's top-left corner, so the translate is the
-      // plain delta between the pane origin and the card origin.
-      transform: `translate(${left}px, ${top - STRIP_HEIGHT_PX}px) scale(${scale})`,
-    };
-  }) as [CardGeometry, CardGeometry];
-
-  return { cards, paneWidth, paneHeight, scale };
-}
-
-type Phase = "idle" | "shrinking" | "zooming";
-
-type StageState = {
-  phase: Phase;
-  from: Mode | null;
-  to: Mode | null;
-  grid: Grid | null;
-};
-
-const IDLE: StageState = { phase: "idle", from: null, to: null, grid: null };
-
-type StageValue = StageState & {
-  switchTo: (from: Mode, to: Mode, href: string) => void;
-};
-
-const ModeSwitchContext = createContext<StageValue>({
-  ...IDLE,
-  switchTo: () => undefined,
-});
-
-export function useModeSwitch(): StageValue {
-  return useContext(ModeSwitchContext);
-}
+import {
+  EASE_OUT,
+  HOLD_MS,
+  IDLE,
+  MODES,
+  ModeSwitchContext,
+  SHRINK_MS,
+  ZOOM_MS,
+  computeGrid,
+  slotOf,
+  useModeSwitch,
+  type Grid,
+  type Mode,
+  type StageState,
+} from "./mode-switch-context";
 
 /**
  * Drives the mode-switch animation. Mounted above the router so it outlives the
